samurai-examples/testcase/testcase/www/html/testcases/css3/flexbox/overflow-keep-scrollpos-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.querySelector('.overflow-auto').scrollTop is 50
